bugfix/BRBDCF-714-janus-won-t-cancel-a-slurm-job : Cancel job allocation for pending slurm state

+++ prov/slurm/executor.go
@@ -17,12 +17,17 @@ import (
 	"novaforge.bull.com/starlings-janus/janus/tosca"
 	"strconv"
 	"strings"
+	"sync"
+	"time"
 )
 
 type defaultExecutor struct {
 	generator defaultGenerator
 	client    *sshutil.SSHClient
 }
+
+const reSallocPending = `^salloc: Pending job allocation (\d+)`
+const reSallocGranted = `^salloc: Granted job allocation (\d+)`
 
 func newExecutor(generator defaultGenerator) prov.DelegateExecutor {
 	return &defaultExecutor{generator: generator}
@@ -129,13 +134,14 @@ func (e *defaultExecutor) uninstallNode(ctx context.Context, kv *api.KV, cfg con
 
 func (e *defaultExecutor) createInfrastructure(ctx context.Context, kv *api.KV, cfg config.Configuration, deploymentID, nodeName string, infra *infrastructure, logOptFields events.LogOptionalFields) error {
 	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString("Creating the slurm infrastructure")
-	var g errgroup.Group
+	g, ctx := errgroup.WithContext(ctx)
+	chAllocationErr := make(chan error)
 	for _, compute := range infra.nodes {
 		func(comp *nodeAllocation) {
 			g.Go(func() error {
-				return e.createNodeAllocation(ctx, comp, deploymentID, nodeName, logOptFields)
+				return e.createNodeAllocation(ctx, comp, deploymentID, nodeName, chAllocationErr, logOptFields)
 			})
-		}(&compute)
+		}(compute)
 	}
 
 	if err := g.Wait(); err != nil {
@@ -145,19 +151,28 @@ func (e *defaultExecutor) createInfrastructure(ctx context.Context, kv *api.KV, 
 		return err
 	}
 
-	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString("Successfully creating the slurm infrastructure")
-	return nil
+	// Handle specific slurm job allocation errors
+	select {
+	case err := <-chAllocationErr:
+		err = errors.Wrapf(err, "Failed to allocate slurm job for deploymentID:%q, node name:%s", deploymentID, nodeName)
+		log.Debugf("%+v", err)
+		events.WithOptionalFields(logOptFields).NewLogEntry(events.ERROR, deploymentID).RegisterAsString(err.Error())
+		return err
+	default:
+		events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString("Successfully creating the slurm infrastructure")
+		return nil
+	}
 }
 
 func (e *defaultExecutor) destroyInfrastructure(ctx context.Context, kv *api.KV, cfg config.Configuration, deploymentID, nodeName string, infra *infrastructure, logOptFields events.LogOptionalFields) error {
 	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString("Destroying the slurm infrastructure")
-	var g errgroup.Group
+	g, ctx := errgroup.WithContext(ctx)
 	for _, compute := range infra.nodes {
 		func(comp *nodeAllocation) {
 			g.Go(func() error {
 				return e.destroyNodeAllocation(ctx, kv, comp, deploymentID, nodeName, logOptFields)
 			})
-		}(&compute)
+		}(compute)
 	}
 
 	if err := g.Wait(); err != nil {
@@ -171,37 +186,68 @@ func (e *defaultExecutor) destroyInfrastructure(ctx context.Context, kv *api.KV,
 	return nil
 }
 
-func (e *defaultExecutor) createNodeAllocation(ctx context.Context, nodeAlloc *nodeAllocation, deploymentID, nodeName string, logOptFields events.LogOptionalFields) error {
+func (e *defaultExecutor) createNodeAllocation(ctx context.Context, nodeAlloc *nodeAllocation, deploymentID, nodeName string, chAllocationErr chan error, logOptFields events.LogOptionalFields) error {
+	log.Debugf("node = %+v", nodeAlloc)
 	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Creating node allocation for: deploymentID:%q, node name:%q", deploymentID, nodeName))
 	// salloc cmd
 	var sallocCPUFlag, sallocMemFlag, sallocPartitionFlag, sallocGresFlag string
 	if nodeAlloc.cpu != "" {
-		sallocCPUFlag = fmt.Sprintf("-c %s", nodeAlloc.cpu)
+		sallocCPUFlag = fmt.Sprintf(" -c %s", nodeAlloc.cpu)
 	}
 	if nodeAlloc.memory != "" {
-		sallocMemFlag = fmt.Sprintf("--mem=%s", nodeAlloc.memory)
+		sallocMemFlag = fmt.Sprintf(" --mem=%s", nodeAlloc.memory)
 	}
 	if nodeAlloc.partition != "" {
-		sallocMemFlag = fmt.Sprintf("-p %s", nodeAlloc.partition)
+		sallocPartitionFlag = fmt.Sprintf(" -p %s", nodeAlloc.partition)
 	}
 	if nodeAlloc.gres != "" {
-		sallocMemFlag = fmt.Sprintf("--gres=%s", nodeAlloc.gres)
+		sallocGresFlag = fmt.Sprintf(" --gres=%s", nodeAlloc.gres)
 	}
 
-	sallocCmd := fmt.Sprintf("salloc --no-shell -J %s %s %s %s %s", nodeAlloc.jobName, sallocCPUFlag, sallocMemFlag, sallocPartitionFlag, sallocGresFlag)
-	sallocOutput, err := e.client.RunCommand(sallocCmd)
+	// salloc command can potentially be a long synchronous command according to the slurm cluster state
+	// so we run it with a session wrapper with stderr/stdout in order to allow job cancellation if user decides to give up the deployment
+	var jobID string
+	var wg sync.WaitGroup
+	sessionWrapper, err := e.client.GetSessionWrapper()
 	if err != nil {
-		return errors.Wrapf(err, "Failed to allocate Slurm resource: %q:", sallocOutput)
+		return errors.Wrap(err, "Failed to get an SSH session wrapper")
 	}
-	// set the jobID
-	split := strings.Split(sallocOutput, " ")
-	jobID := strings.TrimSpace(split[len(split)-1])
-	err = deployments.SetInstanceAttribute(deploymentID, nodeName, nodeAlloc.instanceName, "job_id", jobID)
-	if err != nil {
-		return errors.Wrapf(err, "Failed to set capability attribute (job_id) for node name:%q, instance name:%q", nodeName, nodeAlloc.instanceName)
-	}
-	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Allocating Job ID:%q", jobID))
 
+	chResult := make(chan string, 1)
+	chOut := make(chan bool, 1)
+	chErr := make(chan error)
+	go parseSallocResponse(sessionWrapper.Stderr, chResult, chOut, chErr)
+	go parseSallocResponse(sessionWrapper.Stdout, chResult, chOut, chErr)
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		select {
+		case jobID = <-chResult:
+			log.Debugf("jobID has been received: %q", jobID)
+			deployments.SetInstanceAttribute(deploymentID, nodeName, nodeAlloc.instanceName, "job_id", jobID)
+			events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Allocating Job ID:%q", jobID))
+			return
+		case err := <-chErr:
+			log.Debug(err.Error())
+			events.WithOptionalFields(logOptFields).NewLogEntry(events.ERROR, deploymentID).RegisterAsString(err.Error())
+			chAllocationErr <- err
+			return
+		case <-time.After(5 * time.Second):
+			log.Println("timeout elapsed waiting for jobID parsing after slurm allocation request")
+			events.WithOptionalFields(logOptFields).NewLogEntry(events.ERROR, deploymentID).RegisterAsString("timeout elapsed waiting for jobID parsing after slurm allocation request")
+			chAllocationErr <- err
+			return
+		}
+	}()
+
+	// Run the salloc command
+	sallocCmd := strings.TrimSpace(fmt.Sprintf("salloc --no-shell -J %s%s%s%s%s", nodeAlloc.jobName, sallocCPUFlag, sallocMemFlag, sallocPartitionFlag, sallocGresFlag))
+	err = sessionWrapper.RunCommand(ctx, sallocCmd)
+	if err != nil {
+		return errors.Wrap(err, "Failed to allocate Slurm resource")
+	}
+
+	wg.Wait() // we wait until jobID has been set
 	// run squeue cmd to get slurm node name
 	squeueCmd := fmt.Sprintf("squeue -n %s -j %s --noheader -o \"%%N\"", nodeAlloc.jobName, jobID)
 	slurmNodeName, err := e.client.RunCommand(squeueCmd)
@@ -237,20 +283,22 @@ func (e *defaultExecutor) createNodeAllocation(ctx context.Context, nodeAlloc *n
 }
 
 func (e *defaultExecutor) destroyNodeAllocation(ctx context.Context, kv *api.KV, nodeAlloc *nodeAllocation, deploymentID, nodeName string, logOptFields events.LogOptionalFields) error {
-	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Destroying node allocation for: deploymentID:%q, node name:%q", deploymentID, nodeName))
+	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Destroying node allocation for: deploymentID:%q, node name:%q, instance name:%q", deploymentID, nodeName, nodeAlloc.instanceName))
 	// scancel cmd
 	found, jobID, err := deployments.GetInstanceAttribute(kv, deploymentID, nodeName, nodeAlloc.instanceName, "job_id")
-	if err != nil {
-		return errors.Wrapf(err, "Failed to retrieve Slurm job ID for node name:%s, instance name:%q: %q:", nodeName, nodeAlloc.instanceName)
+	if jobID != "" {
+		if err != nil {
+			return errors.Wrapf(err, "Failed to retrieve Slurm job ID for node name:%s, instance name:%q: %q:", nodeName, nodeAlloc.instanceName)
+		}
+		if !found {
+			return errors.Errorf("Failed to retrieve Slurm job ID for node name:%s, instance name:%q:", nodeName, nodeAlloc.instanceName)
+		}
+		scancelCmd := fmt.Sprintf("scancel %s", jobID)
+		sCancelOutput, err := e.client.RunCommand(scancelCmd)
+		if err != nil {
+			return errors.Wrapf(err, "Failed to cancel Slurm job: %s:", sCancelOutput)
+		}
+		events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Cancelling Job ID:%q", jobID))
 	}
-	if !found {
-		return errors.Errorf("Failed to retrieve Slurm job ID for node name:%s, instance name:%q:", nodeName, nodeAlloc.instanceName)
-	}
-	scancelCmd := fmt.Sprintf("scancel %s", jobID)
-	sCancelOutput, err := e.client.RunCommand(scancelCmd)
-	if err != nil {
-		return errors.Wrapf(err, "Failed to cancel Slurm job: %s:", sCancelOutput)
-	}
-	events.WithOptionalFields(logOptFields).NewLogEntry(events.INFO, deploymentID).RegisterAsString(fmt.Sprintf("Cancelling Job ID:%q", jobID))
 	return nil
 }
